Move opening of facility selection screen to Mobius
