+++ app/src/main/java/org/simple/clinic/registration/location/RegistrationLocationPermissionEffect.kt
@@ -1,3 +1,5 @@
 package org.simple.clinic.registration.location
 
 sealed class RegistrationLocationPermissionEffect
+
+object OpenFacilitySelectionScreen: RegistrationLocationPermissionEffect()
