+++ app/src/main/java/org/simple/clinic/registration/location/RegistrationLocationPermissionEffectHandler.kt
@@ -4,8 +4,10 @@ import com.spotify.mobius.rx2.RxMobius
 import com.squareup.inject.assisted.Assisted
 import com.squareup.inject.assisted.AssistedInject
 import io.reactivex.ObservableTransformer
+import org.simple.clinic.util.scheduler.SchedulersProvider
 
 class RegistrationLocationPermissionEffectHandler @AssistedInject constructor(
+    private val schedulers: SchedulersProvider,
     @Assisted private val uiActions: RegistrationLocationPermissionUiActions
 ) {
 
@@ -17,6 +19,7 @@ class RegistrationLocationPermissionEffectHandler @AssistedInject constructor(
   fun build(): ObservableTransformer<RegistrationLocationPermissionEffect, RegistrationLocationPermissionEvent> {
     return RxMobius
         .subtypeEffectHandler<RegistrationLocationPermissionEffect, RegistrationLocationPermissionEvent>()
+        .addAction(OpenFacilitySelectionScreen::class.java, uiActions::openFacilitySelectionScreen, schedulers.ui())
         .build()
   }
 }
