+++ app/src/main/java/org/simple/clinic/registration/location/RegistrationLocationPermissionUi.kt
@@ -1,5 +1,4 @@
 package org.simple.clinic.registration.location
 
 interface RegistrationLocationPermissionUi: RegistrationLocationPermissionUiActions {
-  fun openFacilitySelectionScreen()
 }
