+++ app/src/main/java/org/simple/clinic/registration/location/RegistrationLocationPermissionUiActions.kt
@@ -1,4 +1,5 @@
 package org.simple.clinic.registration.location
 
 interface RegistrationLocationPermissionUiActions {
+  fun openFacilitySelectionScreen()
 }
