+++ app/src/main/java/org/simple/clinic/registration/location/RegistrationLocationPermissionUpdate.kt
@@ -3,6 +3,7 @@ package org.simple.clinic.registration.location
 import com.spotify.mobius.Next
 import com.spotify.mobius.Next.noChange
 import com.spotify.mobius.Update
+import org.simple.clinic.mobius.dispatch
 
 class RegistrationLocationPermissionUpdate : Update<RegistrationLocationPermissionModel, RegistrationLocationPermissionEvent, RegistrationLocationPermissionEffect> {
 
@@ -10,6 +11,13 @@ class RegistrationLocationPermissionUpdate : Update<RegistrationLocationPermissi
       model: RegistrationLocationPermissionModel,
       event: RegistrationLocationPermissionEvent
   ): Next<RegistrationLocationPermissionModel, RegistrationLocationPermissionEffect> {
-    return noChange()
+    return when (event) {
+      is RequestLocationPermission -> {
+        if (event.isPermissionGranted)
+          dispatch(OpenFacilitySelectionScreen as RegistrationLocationPermissionEffect)
+        else
+          noChange()
+      }
+    }
   }
 }
